Custom controller (visual representation) for Oculus Go
by Joe Strout https://forum.unity.com/members/joestrout.32300/
    from https://forum.unity.com/threads/oculus-go-controller.555949/#post-3681937

Usage:
Attach to some object in scene.

Create whatever you want to represent the controller visually (I typically start in a new project with just a scaled cube, with maybe another cube sticking out of one face like a wand). 

Drag that into this script's controller slot

That's it. Build & run, and you should see your object moving around as you move the physical controller.
Note that we're only changing the local position and rotation with this script. And as it happens, the XR system only changes the local rotation of the camera. Probably you'll want to put the visual controller object, and the main camera, under some other game object that represents the player. Then you can move/rotate this player object, and both the camera and the controller will go along with it.